refactor!: remove `LorentzTransformer` example algorithm
This algorithm was the first transformer-type algorithm, but it is not
so useful and difficult to generalize. Algorithms that need Lorentz
transformations should just do it themselves; if we see certain
transformations being repeated in various algorithms, then we can
re-write such a Lorentz transformation algorithm.

Let's remove this algorithm for now, since we now have other, much more
useful transformer algorithms.