build: use static project version number
We've been using `detect-version.sh` to automatically get the version
number from the most recent `git` tag, but this relies on the `git`
metadata being available at build time. Since this may not always be
true, especially in deployment / release builds, we want to avoid the
situation of deploying `iguana.pc` with an unknown version number. We
thus switch to using a static version number instead.

The only drawback is that we now need to remember to bump the version
number prior to a release.