fix: force usage of off-site RCDB URI
GitHub actions does not respect custom image entrypoints. Our entrypoint
in `container-forge` uses `hostname` to choose between on-site and
off-site database URIs; the default is on-site, since we anticipate most
production usage to be on-site (on `ifarm`).

Since the default is on-site, and GitHub runners are off-site, we force
the usage of the off-site URI.