proof of concept: inherit from another algorithm
The current `FiducialFilter` is for RG-A Pass 1. I am trying to come up
with ways to handle algorithms that are not _quite_ run group
independent. Suppose RG-C would like to use this algorithm, but with a
_different_ configuration, and possibly add some additional cuts. One
approach is to _inherit_ from `FiducialFilter`.

This PR tests this capability by creating an `rgc` namespace, and the
algorithm `rgc::FiducialFilter`.

**do not merge**