build: use `$RCDB_HOME` if option `rcdb:home` unset
If the user does not set build option `rcdb:home`, try `$RCDB_HOME`.
Note that reading environment veriables is strictly forbidden by
`meson`; it would be better for RCDB to be properly packaged.

However, we rely on `$RCDB_HOME` being set in many other consumers, so
we may as well allow it to be used here; this makes builing easier for
most cases. A status message is printed to inform the user when
`$RCDB_HOME` is used at build-time.